Remove unused code now that most Clickhouse functions are on the license server.
