+++ packages/back-end/src/models/DataSourceModel.ts
@@ -28,7 +28,7 @@ import { IS_CLOUD } from "back-end/src/util/secrets";
 import { ReqContext } from "back-end/types/request";
 import { ApiReqContext } from "back-end/types/api";
 import { logger } from "back-end/src/util/logger";
-import { deleteClickhouseUser } from "back-end/src/services/clickhouse";
+import { deleteClickhouseUser } from "back-end/src/services/licenseServerManagedClickhouse";
 import { createModelAuditLogger } from "back-end/src/services/audit";
 import { deleteFactTable, getFactTable } from "./FactTableModel";
 
@@ -57,7 +57,6 @@ const dataSourceSchema = new mongoose.Schema<DataSourceDocument>({
     index: true,
   },
   settings: {},
-  lockUntil: Date,
 });
 dataSourceSchema.index({ id: 1, organization: 1 }, { unique: true });
 type DataSourceDocument = mongoose.Document & DataSourceInterface;
@@ -407,63 +406,6 @@ export async function updateDataSource(
   );
 
   await audit.logUpdate(context, datasource, { ...datasource, ...updates });
-}
-
-function isLocked(datasource: DataSourceInterface): boolean {
-  if (usingFileConfig() || !datasource.lockUntil) return false;
-  return datasource.lockUntil > new Date();
-}
-
-export async function lockDataSource(
-  context: ReqContext | ApiReqContext,
-  datasource: DataSourceInterface,
-  seconds: number,
-) {
-  if (usingFileConfig()) {
-    throw new Error("Cannot lock. Data sources managed by config.yml");
-  }
-  if (datasource.organization !== context.org.id) {
-    throw new Error("Cannot lock data source from another organization");
-  }
-
-  // Already locked, throw error
-  if (isLocked(datasource)) {
-    throw new Error(
-      "Data source is currently being modified. Please try again later.",
-    );
-  }
-
-  await DataSourceModel.updateOne(
-    {
-      id: datasource.id,
-      organization: context.org.id,
-    },
-    {
-      $set: { lockUntil: new Date(Date.now() + seconds * 1000) },
-    },
-  );
-}
-
-export async function unlockDataSource(
-  context: ReqContext | ApiReqContext,
-  datasource: DataSourceInterface,
-) {
-  if (usingFileConfig()) {
-    throw new Error("Cannot unlock. Data sources managed by config.yml");
-  }
-  if (datasource.organization !== context.org.id) {
-    throw new Error("Cannot unlock data source from another organization");
-  }
-
-  await DataSourceModel.updateOne(
-    {
-      id: datasource.id,
-      organization: context.org.id,
-    },
-    {
-      $set: { lockUntil: null },
-    },
-  );
 }
 
 // WARNING: This does not restrict by organization
