+++ packages/back-end/src/services/clickhouse.ts
@@ -1,20 +1,15 @@
-import * as crypto from "crypto";
 import { createClient as createClickhouseClient } from "@clickhouse/client";
-import generator from "generate-password";
 import { AIPromptType } from "shared/ai";
 import { MANAGED_WAREHOUSE_EVENTS_FACT_TABLE_ID } from "shared/constants";
 import {
   isManagedWarehouseAwaitingProvisioning,
   parseIntWithDefault,
 } from "shared/util";
-import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import {
   GrowthbookClickhouseDataSource,
-  DataSourceParams,
   MaterializedColumn,
 } from "shared/types/datasource";
 import { DailyUsage } from "shared/types/organization";
-import { FactTableColumnType } from "shared/types/fact-table";
 import {
   CLICKHOUSE_HOST,
   CLICKHOUSE_ADMIN_USER,
@@ -23,9 +18,7 @@ import {
   CLICKHOUSE_MAIN_TABLE,
   ENVIRONMENT,
   IS_CLOUD,
-  CLICKHOUSE_DEV_PREFIX,
   CLICKHOUSE_OVERAGE_TABLE,
-  MANAGED_CLICKHOUSE_USE_LICENSE_SERVER,
 } from "back-end/src/util/secrets";
 import type { ReqContext } from "back-end/types/request";
 import { logger } from "back-end/src/util/logger";
@@ -33,18 +26,7 @@ import {
   getFactTablesForDatasource,
   updateFactTableColumns,
 } from "back-end/src/models/FactTableModel";
-import {
-  lockDataSource,
-  unlockDataSource,
-} from "back-end/src/models/DataSourceModel";
-import {
-  addCloudSDKMappingViaLicenseServer,
-  createClickhouseUserViaLicenseServer,
-  dangerousRecreateClickhouseTablesViaLicenseServer,
-  deleteClickhouseUserViaLicenseServer,
-  migrateOverageEventsForOrgIdViaLicenseServer,
-  updateMaterializedColumnsInClickhouseViaLicenseServer,
-} from "back-end/src/services/licenseServerManagedClickhouse";
+import { updateMaterializedColumnsInClickhouse } from "back-end/src/services/licenseServerManagedClickhouse";
 
 type ClickHouseDataType =
   | "DateTime"
@@ -53,33 +35,6 @@ type ClickHouseDataType =
   | "String"
   | "LowCardinality(String)";
 
-// These will eventually move to be inside of attributes
-const tempTopLevelFields: Record<string, ClickHouseDataType> = {
-  user_id: "String",
-  url: "String",
-  url_path: "String",
-  url_host: "String",
-  url_query: "String",
-  url_fragment: "String",
-  device_id: "String",
-  page_id: "String",
-  session_id: "String",
-  page_title: "String",
-  utm_source: "String",
-  utm_medium: "String",
-  utm_campaign: "String",
-  utm_term: "String",
-  utm_content: "String",
-  geo_country: "String",
-  geo_city: "String",
-  geo_lat: "Float64",
-  geo_lon: "Float64",
-  ua: "String",
-  ua_browser: "String",
-  ua_os: "String",
-  ua_device_type: "String",
-};
-
 const REMAINING_COLUMNS_SCHEMA: Record<string, ClickHouseDataType> = {
   environment: "LowCardinality(String)",
   sdk_language: "LowCardinality(String)",
@@ -87,17 +42,6 @@ const REMAINING_COLUMNS_SCHEMA: Record<string, ClickHouseDataType> = {
   event_uuid: "String",
   ip: "String",
 };
-
-function clickhouseUserId(orgId: string) {
-  // Sanity check. An orgId of `default` or another reserved word would seriously mess things up
-  if (!orgId.startsWith("org_")) {
-    throw new Error("Invalid organization id");
-  }
-
-  return ENVIRONMENT === "production"
-    ? `${orgId}`
-    : `${CLICKHOUSE_DEV_PREFIX}${orgId}`;
-}
 
 function ensureClickhouseEnvVars() {
   if (
@@ -128,49 +72,6 @@ function createAdminClickhouseClient() {
   });
 }
 
-function getClickhouseDatatype(
-  columnType: FactTableColumnType,
-): ClickHouseDataType {
-  switch (columnType) {
-    case "date":
-      return "DateTime";
-    case "number":
-      return "Float64";
-    case "boolean":
-      return "Boolean";
-    default:
-      return "String";
-  }
-}
-
-function getClickhouseExtractClause(
-  sourceField: string,
-  columnType: FactTableColumnType,
-) {
-  // Some fields will eventually be inside attributes instead of top-level
-  // This is a temp workaround until then
-  if (tempTopLevelFields[sourceField]) {
-    const desiredDataType = getClickhouseDatatype(columnType);
-
-    // If the desired data type is different from the actual type, need to cast it
-    if (desiredDataType !== tempTopLevelFields[sourceField]) {
-      return `CAST(${sourceField} AS ${desiredDataType})`;
-    }
-
-    // Otherwise, just return the column name
-    return sourceField;
-  }
-
-  switch (columnType) {
-    case "number":
-      return `JSONExtractFloat(context_json, '${sourceField}')`;
-    case "boolean":
-      return `JSONExtractBool(context_json, '${sourceField}')`;
-    default:
-      return `JSONExtractString(context_json, '${sourceField}')`;
-  }
-}
-
 export function getReservedColumnNames(): Set<string> {
   return new Set(
     [
@@ -183,388 +84,6 @@ export function getReservedColumnNames(): Set<string> {
       "variation_id",
       ...Object.keys(REMAINING_COLUMNS_SCHEMA),
     ].map((col) => col.toLowerCase()),
-  );
-}
-
-type ColumnDef = {
-  source: string;
-  alias?: string;
-  datatype: ClickHouseDataType;
-};
-
-function getCreateTableColumnList(columns: ColumnDef[]): string[] {
-  return columns.map(
-    ({ source, alias, datatype }) => `${alias || source} ${datatype}`,
-  );
-}
-function getSelectColumnList(columns: ColumnDef[]): string[] {
-  return columns.map(
-    ({ source, alias }) =>
-      `${source}${alias && alias !== source ? ` as ${alias}` : ""}`,
-  );
-}
-
-function getRemainingColumnDefs(): ColumnDef[] {
-  return Object.entries(REMAINING_COLUMNS_SCHEMA).map(([colName, colType]) => ({
-    source: colName,
-    datatype: colType as ClickHouseDataType,
-  }));
-}
-
-function getMaterializedColumnDefs(
-  materializedColumns: MaterializedColumn[],
-): ColumnDef[] {
-  return materializedColumns.map(({ columnName, datatype, sourceField }) => ({
-    source: getClickhouseExtractClause(sourceField, datatype),
-    alias: columnName,
-    datatype: getClickhouseDatatype(datatype),
-  }));
-}
-
-function getMaterializedViewSQL({
-  orgId,
-  colDefs,
-  orderBy,
-  filter,
-  baseTableName,
-}: {
-  orgId: string;
-  colDefs: ColumnDef[];
-  orderBy: string;
-  filter: string;
-  baseTableName: string;
-}): {
-  createTable: string;
-  createView: string;
-  populateTable: string;
-  select: string;
-  tableName: string;
-  viewName: string;
-} {
-  const tableName = getTableName(orgId, baseTableName);
-  const viewName = getTableName(orgId, `${baseTableName}_mv`);
-
-  const createTable = `CREATE TABLE ${tableName} (
-  ${getCreateTableColumnList(colDefs).join(",\n  ")}
-) ENGINE = MergeTree()
-PARTITION BY toYYYYMM(timestamp) 
-ORDER BY ${orderBy}`;
-
-  const select = `SELECT ${getSelectColumnList(colDefs).join(", ")}
-    FROM ${CLICKHOUSE_MAIN_TABLE} 
-    WHERE (organization = '${orgId}') AND (${filter})`;
-
-  const populateTable = `INSERT INTO ${tableName} ${select}`;
-  const createView = `CREATE MATERIALIZED VIEW ${viewName} TO ${tableName} 
-DEFINER=CURRENT_USER SQL SECURITY DEFINER
-AS ${select}`;
-
-  return {
-    createTable,
-    select,
-    populateTable,
-    createView,
-    tableName,
-    viewName,
-  };
-}
-
-function getEventsSQL(
-  orgId: string,
-  materializedColumns: MaterializedColumn[],
-) {
-  return getMaterializedViewSQL({
-    orgId,
-    baseTableName: "events",
-    filter: "event_name NOT IN ('Experiment Viewed', 'Feature Evaluated')",
-    orderBy: "(event_name, timestamp)",
-    colDefs: [
-      { source: "timestamp", datatype: "DateTime" },
-      { source: "client_key", datatype: "String" },
-      { source: "event_name", datatype: "String" },
-      { source: "properties_json", alias: "properties", datatype: "String" },
-      { source: "context_json", alias: "attributes", datatype: "String" },
-      ...getRemainingColumnDefs(),
-      ...getMaterializedColumnDefs(materializedColumns),
-    ],
-  });
-}
-
-function getExperimentViewSQL(
-  orgId: string,
-  materializedColumns: MaterializedColumn[],
-) {
-  return getMaterializedViewSQL({
-    orgId,
-    baseTableName: "experiment_views",
-    filter: "event_name = 'Experiment Viewed'",
-    orderBy: "(experiment_id, timestamp)",
-    colDefs: [
-      { source: "timestamp", datatype: "DateTime" },
-      { source: "client_key", datatype: "String" },
-      {
-        source: "JSONExtractString(properties_json, 'experimentId')",
-        alias: "experiment_id",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'variationId')",
-        alias: "variation_id",
-        datatype: "String",
-      },
-      { source: "properties_json", alias: "properties", datatype: "String" },
-      { source: "context_json", alias: "attributes", datatype: "String" },
-      ...getRemainingColumnDefs(),
-      ...getMaterializedColumnDefs(materializedColumns),
-    ],
-  });
-}
-
-function getFeatureusageSQL(orgId: string) {
-  return getMaterializedViewSQL({
-    orgId,
-    baseTableName: "feature_usage",
-    filter: "event_name = 'Feature Evaluated'",
-    orderBy: "(feature, timestamp)",
-    colDefs: [
-      { source: "timestamp", datatype: "DateTime" },
-      { source: "client_key", datatype: "String" },
-      {
-        source: "JSONExtractString(properties_json, 'feature')",
-        alias: "feature",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'revision')",
-        alias: "revision",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'source')",
-        alias: "source",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'value')",
-        alias: "value",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'ruleId')",
-        alias: "ruleId",
-        datatype: "String",
-      },
-      {
-        source: "JSONExtractString(properties_json, 'variationId')",
-        alias: "variationId",
-        datatype: "String",
-      },
-      { source: "context_json", alias: "attributes", datatype: "String" },
-      ...getRemainingColumnDefs(),
-    ],
-  });
-}
-
-async function runCommand(
-  client: ReturnType<typeof createClickhouseClient>,
-  query: string,
-): Promise<void> {
-  await client.command({ query });
-}
-
-function getTableName(orgId: string, name: string) {
-  const user = clickhouseUserId(orgId);
-  const database = user;
-  return `${database}.${name}`;
-}
-
-export async function createClickhouseUser(
-  context: ReqContext,
-  materializedColumns: MaterializedColumn[] = [],
-): Promise<DataSourceParams> {
-  if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-    return createClickhouseUserViaLicenseServer(
-      context.org.id,
-      materializedColumns,
-    );
-  }
-
-  const client = createAdminClickhouseClient();
-
-  const orgId = context.org.id;
-  const user = clickhouseUserId(orgId);
-  const password = generator.generate({
-    length: 30,
-    numbers: true,
-  });
-  const hashedPassword = crypto
-    .createHash("sha256")
-    .update(password)
-    .digest("hex");
-
-  const database = user;
-  logger.info(`creating Clickhouse database ${database}`);
-  // It's important this does not have "IF NOT EXISTS" to protect against race conditions
-  await runCommand(client, `CREATE DATABASE ${database}`);
-
-  logger.info(`Creating Clickhouse user ${user}`);
-  await runCommand(
-    client,
-    `CREATE USER ${user} IDENTIFIED WITH sha256_hash BY '${hashedPassword}' DEFAULT DATABASE ${database}`,
-  );
-
-  await createClickhouseTables(client, orgId, materializedColumns);
-
-  logger.info(
-    `Granting select permissions on information_schema.columns to ${user}`,
-  );
-  // For schema browser.  They can only see info on tables that they have select permissions on.
-  await runCommand(
-    client,
-    `GRANT SELECT(data_type, table_name, table_catalog, table_schema, column_name) ON information_schema.columns TO ${user}`,
-  );
-
-  const url = new URL(CLICKHOUSE_HOST);
-
-  const params = {
-    port: parseIntWithDefault(url.port, 9000),
-    url: url.toString(),
-    user: user,
-    password: password,
-    database: database,
-  };
-
-  return params;
-}
-
-export async function createClickhouseTables(
-  client: ReturnType<typeof createAdminClickhouseClient>,
-  orgId: string,
-  materializedColumns: MaterializedColumn[] = [],
-): Promise<void> {
-  const user = clickhouseUserId(orgId);
-  const database = user;
-
-  // Events table
-  const eventsSQL = getEventsSQL(orgId, materializedColumns);
-  logger.info(`Creating table ${eventsSQL.tableName}`);
-  await runCommand(client, eventsSQL.createTable);
-  logger.info(`Populating table ${eventsSQL.tableName}`);
-  await runCommand(client, eventsSQL.populateTable);
-  logger.info(`Creating materialized view ${eventsSQL.viewName}`);
-  await runCommand(client, eventsSQL.createView);
-
-  // Experiment views table
-  const experimentViewSQL = getExperimentViewSQL(orgId, materializedColumns);
-  logger.info(`Creating table ${experimentViewSQL.tableName}`);
-  await runCommand(client, experimentViewSQL.createTable);
-  logger.info(`Populating table ${experimentViewSQL.tableName}`);
-  await runCommand(client, experimentViewSQL.populateTable);
-  logger.info(`Creating materialized view ${experimentViewSQL.viewName}`);
-  await runCommand(client, experimentViewSQL.createView);
-
-  // Feature usage table
-  const featureUsageSQL = getFeatureusageSQL(orgId);
-  logger.info(`Creating table ${featureUsageSQL.tableName}`);
-  await runCommand(client, featureUsageSQL.createTable);
-  logger.info(`Populating table ${featureUsageSQL.tableName}`);
-  await runCommand(client, featureUsageSQL.populateTable);
-  logger.info(`Creating materialized view ${featureUsageSQL.viewName}`);
-  await runCommand(client, featureUsageSQL.createView);
-
-  logger.info(`Granting select permissions on ${database}.* to ${user}`);
-  await runCommand(client, `GRANT SELECT ON ${database}.* TO ${user}`);
-}
-
-export async function _dangerousRecreateClickhouseTables(
-  context: ReqContext,
-  datasource: GrowthbookClickhouseDataSource,
-): Promise<void> {
-  const orgId = context.org.id;
-
-  try {
-    if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-      await dangerousRecreateClickhouseTablesViaLicenseServer(
-        orgId,
-        datasource.settings.materializedColumns || [],
-      );
-    } else {
-      // Backfilling data can take a while, so lock the datasource for 30 minutes
-      await lockDataSource(context, datasource, 1800);
-
-      const client = createAdminClickhouseClient();
-      const user = clickhouseUserId(orgId);
-      const database = user;
-
-      // Drop the entire database and recreate it
-      logger.info(`Dropping Clickhouse database ${database}`);
-      await runCommand(client, `DROP DATABASE IF EXISTS ${database}`);
-
-      logger.info(`Creating Clickhouse database ${database}`);
-      await runCommand(client, `CREATE DATABASE ${database}`);
-      await createClickhouseTables(
-        client,
-        orgId,
-        datasource.settings.materializedColumns || [],
-      );
-    }
-  } finally {
-    await unlockDataSource(context, datasource);
-  }
-}
-
-export async function deleteClickhouseUser(organization: string) {
-  if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-    return deleteClickhouseUserViaLicenseServer(organization);
-  }
-
-  const client = createAdminClickhouseClient();
-  const user = clickhouseUserId(organization);
-  const database = user;
-
-  logger.info(`Deleting Clickhouse user ${user}`);
-  await runCommand(client, `DROP USER IF EXISTS ${user}`);
-
-  logger.info(`Deleting Clickhouse database ${database}`);
-  await runCommand(client, `DROP DATABASE IF EXISTS ${database}`);
-}
-
-export async function addCloudSDKMapping(connection: SDKConnectionInterface) {
-  const { key, organization } = connection;
-
-  // This is not a fatal error, so just log instead of throwing
-  try {
-    if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-      await addCloudSDKMappingViaLicenseServer(key, organization);
-    } else {
-      const client = createAdminClickhouseClient();
-      await client.insert({
-        table: "usage.sdk_key_mapping",
-        values: [{ key, organization }],
-        format: "JSONEachRow",
-      });
-    }
-  } catch (e) {
-    logger.error(
-      e,
-      `Error inserting sdk key mapping (${key} -> ${organization})`,
-    );
-  }
-}
-
-export async function migrateOverageEventsForOrgId(orgId: string) {
-  if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-    return migrateOverageEventsForOrgIdViaLicenseServer(orgId);
-  }
-
-  const client = createAdminClickhouseClient();
-  await runCommand(
-    client,
-    `INSERT INTO ${CLICKHOUSE_MAIN_TABLE} SELECT * FROM ${CLICKHOUSE_OVERAGE_TABLE} WHERE organization = '${orgId}'`,
-  );
-  await runCommand(
-    client,
-    `ALTER TABLE ${CLICKHOUSE_OVERAGE_TABLE} DELETE WHERE organization = '${orgId}'`,
   );
 }
 
@@ -734,101 +253,14 @@ export async function updateMaterializedColumns({
   }
   const orgId = datasource.organization;
 
-  try {
-    if (MANAGED_CLICKHOUSE_USE_LICENSE_SERVER) {
-      await updateMaterializedColumnsInClickhouseViaLicenseServer({
-        orgId,
-        columnsToAdd,
-        columnsToDelete,
-        columnsToRename,
-        finalColumns,
-        originalColumns,
-      });
-    } else {
-      // We can only process one materialized column update at a time.
-      // This should be quick, but lock it 5 minutes just in case.
-      await lockDataSource(context, datasource, 300);
-
-      const client = createAdminClickhouseClient();
-
-      const addClauses = columnsToAdd
-        .map(
-          ({ columnName, datatype }) =>
-            `ADD COLUMN IF NOT EXISTS ${columnName} ${getClickhouseDatatype(
-              datatype,
-            )}`,
-        )
-        .join(", ");
-      const dropClauses = columnsToDelete
-        .map((columnName) => `DROP COLUMN IF EXISTS ${columnName}`)
-        .join(", ");
-      const renameClauses = columnsToRename
-        .map(({ from, to }) => `RENAME COLUMN ${from} to ${to}`)
-        .join(", ");
-      const clauses = `${addClauses}${
-        columnsToAdd.length > 0 &&
-        columnsToDelete.length + columnsToRename.length > 0
-          ? ", "
-          : ""
-      }${dropClauses}${
-        columnsToDelete.length > 0 && columnsToRename.length > 0 ? ", " : ""
-      }${renameClauses}`;
-
-      // Track which columns the view should be recreated with in case of an error
-      let viewColumns = originalColumns;
-
-      // First update the main events table
-      const { tableName: eventsTableName, viewName: eventsViewName } =
-        getEventsSQL(orgId, []);
-      logger.info(
-        `Updating materialized columns; dropping view ${eventsViewName}`,
-      );
-      await runCommand(client, `DROP VIEW IF EXISTS ${eventsViewName}`);
-      let err = undefined;
-      try {
-        logger.info(`Updating table schema for ${eventsTableName}`);
-        await runCommand(client, `ALTER TABLE ${eventsTableName} ${clauses}`);
-        viewColumns = finalColumns;
-      } catch (e) {
-        logger.error(e);
-        err = e;
-      } finally {
-        logger.info(`Recreating materialized view ${eventsViewName}`);
-        const eventsSQL = getEventsSQL(orgId, viewColumns);
-        await runCommand(client, eventsSQL.createView);
-      }
-      if (err) {
-        throw err;
-      }
-
-      // Now update the experiment views table
-      const { tableName: exposureTableName, viewName: exposureViewName } =
-        getExperimentViewSQL(orgId, []);
-      logger.info(
-        `Updating materialized columns; dropping view ${exposureViewName}`,
-      );
-      await runCommand(client, `DROP VIEW IF EXISTS ${exposureViewName}`);
-      err = undefined;
-      viewColumns = originalColumns;
-      try {
-        logger.info(`Updating table schema for ${exposureTableName}`);
-        await runCommand(client, `ALTER TABLE ${exposureTableName} ${clauses}`);
-        viewColumns = finalColumns;
-      } catch (e) {
-        logger.error(e);
-        err = e;
-      } finally {
-        logger.info(`Recreating materialized view ${exposureViewName}`);
-        const experimentViewSQL = getExperimentViewSQL(orgId, viewColumns);
-        await runCommand(client, experimentViewSQL.createView);
-      }
-      if (err) {
-        throw err;
-      }
-    }
-  } finally {
-    await unlockDataSource(context, datasource);
-  }
+  await updateMaterializedColumnsInClickhouse({
+    orgId,
+    columnsToAdd,
+    columnsToDelete,
+    columnsToRename,
+    finalColumns,
+    originalColumns,
+  });
 
   // Update the main events fact table with the new columns
   const factTables = await getFactTablesForDatasource(context, datasource.id);
