+++ packages/back-end/src/util/secrets.ts
@@ -325,14 +325,6 @@ export const CLICKHOUSE_OVERAGE_TABLE =
 export const CLICKHOUSE_DEV_PREFIX =
   process.env.CLICKHOUSE_DEV_PREFIX || "test_";
 
-/** When true, managed warehouse ClickHouse provisioning runs on central-license-server.
- * TODO(james): remove this once we are sure we don't need to rollback.
- */
-export const MANAGED_CLICKHOUSE_USE_LICENSE_SERVER = stringToBoolean(
-  process.env.MANAGED_CLICKHOUSE_USE_LICENSE_SERVER,
-  IS_CLOUD,
-);
-
 export const CLOUD_SECRET = process.env.CLOUD_SECRET ?? "";
 
 // Note: the Visual Editor relies on the information in this path, so disabling it will prevent some features from working correctly.
