+++ packages/back-end/test/services/clickhouse.test.ts
@@ -6,14 +6,11 @@ import {
 import { MANAGED_WAREHOUSE_EVENTS_FACT_TABLE_ID } from "shared/constants";
 import type { ReqContext } from "back-end/types/request";
 import { updateMaterializedColumns } from "back-end/src/services/clickhouse";
+import { updateMaterializedColumnsInClickhouse } from "back-end/src/services/licenseServerManagedClickhouse";
 import {
   getFactTablesForDatasource,
   updateFactTableColumns,
 } from "back-end/src/models/FactTableModel";
-import {
-  lockDataSource,
-  unlockDataSource,
-} from "back-end/src/models/DataSourceModel";
 
 const mockCommand = jest.fn();
 const mockClickhouseClient = {
@@ -34,7 +31,10 @@ jest.mock("back-end/src/util/secrets", () => ({
   IS_CLOUD: false,
   CLICKHOUSE_DEV_PREFIX: "dev_",
   CLICKHOUSE_OVERAGE_TABLE: "overage",
-  MANAGED_CLICKHOUSE_USE_LICENSE_SERVER: false,
+}));
+
+jest.mock("back-end/src/services/licenseServerManagedClickhouse", () => ({
+  updateMaterializedColumnsInClickhouse: jest.fn().mockResolvedValue(undefined),
 }));
 
 jest.mock("back-end/src/models/FactTableModel", () => ({
@@ -42,15 +42,11 @@ jest.mock("back-end/src/models/FactTableModel", () => ({
   updateFactTableColumns: jest.fn(),
 }));
 
-jest.mock("back-end/src/models/DataSourceModel", () => ({
-  lockDataSource: jest.fn(),
-  unlockDataSource: jest.fn(),
-}));
-
 const mockGetFactTablesForDatasource = jest.mocked(getFactTablesForDatasource);
 const mockUpdateFactTableColumns = jest.mocked(updateFactTableColumns);
-const mockLockDataSource = jest.mocked(lockDataSource);
-const mockUnlockDataSource = jest.mocked(unlockDataSource);
+const mockUpdateMaterializedColumnsLicense = jest.mocked(
+  updateMaterializedColumnsInClickhouse,
+);
 
 function makeFactTableColumn(
   column: string,
@@ -92,8 +88,6 @@ describe("updateMaterializedColumns", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     mockCommand.mockResolvedValue(undefined);
-    mockLockDataSource.mockResolvedValue(undefined as never);
-    mockUnlockDataSource.mockResolvedValue(undefined as never);
     mockUpdateFactTableColumns.mockResolvedValue(undefined as never);
   });
 
@@ -124,6 +118,7 @@ describe("updateMaterializedColumns", () => {
       originalColumns: [],
     });
 
+    expect(mockUpdateMaterializedColumnsLicense).toHaveBeenCalledTimes(1);
     expect(mockUpdateFactTableColumns).toHaveBeenCalledTimes(1);
     const changes = mockUpdateFactTableColumns.mock.calls[0][1] as {
       columns: ColumnInterface[];
@@ -135,7 +130,7 @@ describe("updateMaterializedColumns", () => {
     );
   });
 
-  it("does not call ClickHouse when managed warehouse has not been provisioned yet", async () => {
+  it("does not call the managed warehouse service when provisioning is not complete", async () => {
     const unprovisionedDs = {
       ...datasource,
       settings: { hasBeenProvisioned: false },
@@ -151,8 +146,8 @@ describe("updateMaterializedColumns", () => {
       originalColumns: [],
     });
 
-    expect(mockLockDataSource).not.toHaveBeenCalled();
     expect(mockCommand).not.toHaveBeenCalled();
+    expect(mockUpdateMaterializedColumnsLicense).not.toHaveBeenCalled();
   });
 
   it("restores deleted rename destination and tombstones source when destination name already exists", async () => {
@@ -183,6 +178,7 @@ describe("updateMaterializedColumns", () => {
       originalColumns: [],
     });
 
+    expect(mockUpdateMaterializedColumnsLicense).toHaveBeenCalledTimes(1);
     expect(mockUpdateFactTableColumns).toHaveBeenCalledTimes(1);
     const changes = mockUpdateFactTableColumns.mock.calls[0][1] as {
       columns: ColumnInterface[];
